pyproject.toml
[project]
name = "pin-people"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "psycopg>=3.3.2",
    "python-dotenv>=1.0",
    "drf-spectacular>=0.27",
]

[dependency-groups]
dev = [
    "blue>=0.9.1",
]
